refactor(crypto): rename module
